Create WhyITookITM352.md
I decided to take ITM 352 because I am very motivated to become a programmer and my home campus does not offer a class specifically on programming application systems. This is my last semester of undergrad and I feel the skills taught in this class will be very helpful to me, professionally, in the future.